fixing the windows build issue
- Fixed the issue with the windows build
- Added a strategy to test all build targets during PR
