+++ src/catalog/cacheservice/main.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/Parallels/prl-devops-service/basecontext"
@@ -148,21 +147,6 @@ func (cs *CacheService) cachedMetadataFilePath() string {
 
 func (cs *CacheService) notify(message string) {
 	cs.notifications.NotifyInfo(message)
-}
-
-func (cs *CacheService) getFreeDiskSpace() (int64, error) {
-	// We will be getting the free disk space from the disk
-	var stat syscall.Statfs_t
-	err := syscall.Statfs(cs.cacheFolder, &stat)
-	if err != nil {
-		return 0, errors.NewFromErrorWithCode(err, 500)
-	}
-
-	// Available blocks * size per block = available space in bytes
-	diskFreeSpaceInBytes := int64(stat.Bavail) * int64(stat.Bsize)
-	diskFreeSpaceInMb := diskFreeSpaceInBytes / (1024 * 1024)
-
-	return diskFreeSpaceInMb, nil
 }
 
 func (cs *CacheService) getCacheTotalSize() (int64, error) {
